feat(mainwindow): Add column/line number and selection count
This feature is requested in coder3101#157.

The implementation follows Sublime Text.

Line number and column number are displayed if there isn't a selection.

Both line number and column number are 1-indexed, i.e. the left-top
corner is line 1 column 1.

The column number takes tabs as special characters, and the result is
the "read" column number, for example:

("|" for the cursor, suppose the tab length is four)

text: `[tab]a|b`
column: 6

text: `[space][tab][space][space][tab][space]|a`
column: 10

Selection character count is displayed if there is a selection.

"1 characters selected" will be displayed instead of "1 character
selected", not only because it's slightly easier to implement, but also
because it will not cause significant UI change when change the
selection length from 1 to 2 or vice versa.

If the selection spreads more than one line, the line count of selection
will also be displayed.